GetLocation OK.

+++ models/request_model.go
@@ -6,6 +6,6 @@ type Request struct {
 
 type RequestSplit struct {
 	Name     string   `uri:"satellite_name"`
-	Distance float32  `json:"distance"`
+	Distance float64  `json:"distance"`
 	Message  []string `json:"message"`
 }
